ContrastAndBrightnessTab: add auto ajust

+++ src/main/java/ksn/imgusage/type/dto/opencv/ContrastAndBrightnessTabParams.java
@@ -8,13 +8,13 @@ import ksn.imgusage.tabs.opencv.ContrastAndBrightnessTab;
 /** Init parameters for {@link ContrastAndBrightnessTab} */
 public class ContrastAndBrightnessTabParams implements ITabParams {
 
-    public double  alpha = 1.0;
-    public int     beta  = 0;
+    public double alpha = 1.0;
+    public double beta  = 0;
 
     @Override
     public String toString() {
         return String.format(Locale.US,
-            "{ alpha=%.2f, beta=%d }",
+            "{ alpha=%.2f, beta=%.2f }",
             alpha,
             beta);
     }
